refactor: Maven executor uses project's pool
