+++ it/src/test/java/org/eclipse/jkube/integrationtests/assertions/PodAssertion.java
@@ -28,7 +28,7 @@ import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 
-import static org.eclipse.jkube.integrationtests.WaitUtil.await;
+import static org.eclipse.jkube.integrationtests.AsyncUtil.await;
 import static org.eclipse.jkube.integrationtests.assertions.LabelAssertion.assertLabels;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.notNullValue;
